requirements.txt
tk-tools
opencv-contrib-python
datetime
pytest-shutil
python-csv
numpy
pillow
pandas
times
mysql-connector-python 